# perf: use GetByValue to avoid boxing to interface{}