--- a/v1/ast/check.go
+++ b/v1/ast/check.go
@@ -276,7 +276,7 @@ func (tc *typeChecker) checkRule(env *TypeEnv, as *AnnotationSet, rule *Rule) {
 	if len(rule.Head.Args) > 0 {
 		// If args are not referred to in body, infer as any.
 		WalkVars(rule.Head.Args, func(v Var) bool {
-			if cpy.Get(v) == nil {
+			if cpy.GetByValue(v) == nil {
 				cpy.tree.PutOne(v, types.A)
 			}
 			return false
@@ -285,7 +285,7 @@ func (tc *typeChecker) checkRule(env *TypeEnv, as *AnnotationSet, rule *Rule) {
 		// Construct function type.
 		args := make([]types.Type, len(rule.Head.Args))
 		for i := range len(rule.Head.Args) {
-			args[i] = cpy.Get(rule.Head.Args[i])
+			args[i] = cpy.GetByValue(rule.Head.Args[i].Value)
 		}
 
 		f := types.NewFunction(args, cpy.Get(rule.Head.Value))
@@ -294,7 +294,7 @@ func (tc *typeChecker) checkRule(env *TypeEnv, as *AnnotationSet, rule *Rule) {
 	} else {
 		switch rule.Head.RuleKind() {
 		case SingleValue:
-			typeV := cpy.Get(rule.Head.Value)
+			typeV := cpy.GetByValue(rule.Head.Value.Value)
 			if !path.IsGround() {
 				// e.g. store object[string: whatever] at data.p.q.r, not data.p.q.r[x] or data.p.q.r[x].y[z]
 				objPath := path.DynamicSuffix()
@@ -312,7 +312,7 @@ func (tc *typeChecker) checkRule(env *TypeEnv, as *AnnotationSet, rule *Rule) {
 				}
 			}
 		case MultiValue:
-			typeK := cpy.Get(rule.Head.Key)
+			typeK := cpy.GetByValue(rule.Head.Key.Value)
 			if typeK != nil {
 				tpe = types.NewSet(typeK)
 			}
@@ -341,7 +341,7 @@ func nestedObject(env *TypeEnv, path Ref, tpe types.Type) (types.Type, error) {
 	}
 
 	var dynamicProperty *types.DynamicProperty
-	typeK := env.Get(k)
+	typeK := env.GetByValue(k.Value)
 	if typeK == nil {
 		return nil, nil
 	}
@@ -391,7 +391,7 @@ func (tc *typeChecker) checkExprBuiltin(env *TypeEnv, expr *Expr) *Error {
 	// type checker relies on reordering (in particular for references to local
 	// vars).
 	name := expr.Operator()
-	tpe := env.Get(name)
+	tpe := env.GetByRef(name)
 
 	if tpe == nil {
 		if tc.allowUndefinedFuncs {
@@ -431,7 +431,7 @@ func (tc *typeChecker) checkExprBuiltin(env *TypeEnv, expr *Expr) *Error {
 		if !unify1(env, args[i], fargs.Arg(i), false) {
 			post := make([]types.Type, len(args))
 			for i := range args {
-				post[i] = env.Get(args[i])
+				post[i] = env.GetByValue(args[i].Value)
 			}
 			return newArgError(expr.Location, name, "invalid argument(s)", post, namedFargs)
 		}
@@ -453,7 +453,7 @@ func checkExprEq(env *TypeEnv, expr *Expr) *Error {
 	}
 
 	a, b := expr.Operand(0), expr.Operand(1)
-	typeA, typeB := env.Get(a), env.Get(b)
+	typeA, typeB := env.GetByValue(a.Value), env.GetByValue(b.Value)
 
 	if !unify2(env, a, typeA, b, typeB) {
 		err := NewError(TypeErr, expr.Location, "match error")
@@ -473,7 +473,7 @@ func (tc *typeChecker) checkExprWith(env *TypeEnv, expr *Expr, i int) *Error {
 	}
 
 	target, value := expr.With[i].Target, expr.With[i].Value
-	targetType, valueType := env.Get(target), env.Get(value)
+	targetType, valueType := env.GetByValue(target.Value), env.GetByValue(value.Value)
 
 	if t, ok := targetType.(*types.Function); ok { // built-in function replacement
 		switch v := valueType.(type) {
@@ -509,7 +509,7 @@ func unify2(env *TypeEnv, a *Term, typeA types.Type, b *Term, typeB types.Type) 
 	case Var:
 		switch b.Value.(type) {
 		case Var:
-			return unify1(env, a, types.A, false) && unify1(env, b, env.Get(a), false)
+			return unify1(env, a, types.A, false) && unify1(env, b, env.GetByValue(a.Value), false)
 		case *Array:
 			return unify2Array(env, b, a)
 		case *object:
@@ -526,14 +526,14 @@ func unify2Array(env *TypeEnv, a *Term, b *Term) bool {
 	case *Array:
 		if arr.Len() == bv.Len() {
 			for i := range arr.Len() {
-				if !unify2(env, arr.Elem(i), env.Get(arr.Elem(i)), bv.Elem(i), env.Get(bv.Elem(i))) {
+				if !unify2(env, arr.Elem(i), env.GetByValue(arr.Elem(i).Value), bv.Elem(i), env.GetByValue(bv.Elem(i).Value)) {
 					return false
 				}
 			}
 			return true
 		}
 	case Var:
-		return unify1(env, a, types.A, false) && unify1(env, b, env.Get(a), false)
+		return unify1(env, a, types.A, false) && unify1(env, b, env.GetByValue(a.Value), false)
 	}
 	return false
 }
@@ -545,14 +545,14 @@ func unify2Object(env *TypeEnv, a *Term, b *Term) bool {
 		cv := obj.Intersect(bv)
 		if obj.Len() == bv.Len() && bv.Len() == len(cv) {
 			for i := range cv {
-				if !unify2(env, cv[i][1], env.Get(cv[i][1]), cv[i][2], env.Get(cv[i][2])) {
+				if !unify2(env, cv[i][1], env.GetByValue(cv[i][1].Value), cv[i][2], env.GetByValue(cv[i][2].Value)) {
 					return false
 				}
 			}
 			return true
 		}
 	case Var:
-		return unify1(env, a, types.A, false) && unify1(env, b, env.Get(a), false)
+		return unify1(env, a, types.A, false) && unify1(env, b, env.GetByValue(a.Value), false)
 	}
 	return false
 }
@@ -615,22 +615,22 @@ func unify1(env *TypeEnv, term *Term, tpe types.Type, union bool) bool {
 		}
 		return false
 	case Ref, *ArrayComprehension, *ObjectComprehension, *SetComprehension:
-		return unifies(env.Get(v), tpe)
+		return unifies(env.GetByValue(v), tpe)
 	case Var:
 		if !union {
-			if exist := env.Get(v); exist != nil {
+			if exist := env.GetByValue(v); exist != nil {
 				return unifies(exist, tpe)
 			}
 			env.tree.PutOne(term.Value, tpe)
 		} else {
-			env.tree.PutOne(term.Value, types.Or(env.Get(v), tpe))
+			env.tree.PutOne(term.Value, types.Or(env.GetByValue(v), tpe))
 		}
 		return true
 	default:
 		if !IsConstant(v) {
 			panic("unreachable")
 		}
-		return unifies(env.Get(term), tpe)
+		return unifies(env.GetByValue(term.Value), tpe)
 	}
 }
 
@@ -732,7 +732,7 @@ func (rc *refChecker) Visit(x interface{}) bool {
 }
 
 func (rc *refChecker) checkApply(curr *TypeEnv, ref Ref) *Error {
-	switch tpe := curr.Get(ref).(type) {
+	switch tpe := curr.GetByRef(ref).(type) {
 	case *types.Function: // NOTE(sr): We don't support first-class functions, except for `with`.
 		return newRefErrUnsupported(ref[0].Location, rc.varRewriter(ref), len(ref)-1, tpe)
 	}
@@ -755,19 +755,19 @@ func (rc *refChecker) checkRef(curr *TypeEnv, node *typeTreeNode, ref Ref, idx i
 		switch head.Value.(type) {
 		case Var, String: // OK
 		default:
-			have := rc.env.Get(head.Value)
+			have := rc.env.GetByValue(head.Value)
 			return newRefErrInvalid(ref[0].Location, rc.varRewriter(ref), idx, have, types.S, getOneOfForNode(node))
 		}
 	}
 
-	if v, ok := head.Value.(Var); ok && idx != 0 {
+	if _, ok := head.Value.(Var); ok && idx != 0 {
 		tpe := types.Keys(rc.env.getRefRecExtent(node))
-		if exist := rc.env.Get(v); exist != nil {
+		if exist := rc.env.GetByValue(head.Value); exist != nil {
 			if !unifies(tpe, exist) {
 				return newRefErrInvalid(ref[0].Location, rc.varRewriter(ref), idx, exist, tpe, getOneOfForNode(node))
 			}
 		} else {
-			rc.env.tree.PutOne(v, tpe)
+			rc.env.tree.PutOne(head.Value, tpe)
 		}
 	}
 
@@ -817,7 +817,7 @@ func (rc *refChecker) checkRefLeaf(tpe types.Type, ref Ref, idx int) *Error {
 	switch value := head.Value.(type) {
 
 	case Var:
-		if exist := rc.env.Get(value); exist != nil {
+		if exist := rc.env.GetByValue(value); exist != nil {
 			if !unifies(exist, keys) {
 				return newRefErrInvalid(ref[0].Location, rc.varRewriter(ref), idx, exist, keys, getOneOfForType(tpe))
 			}