--- a/v1/ast/env.go
+++ b/v1/ast/env.go
@@ -29,29 +29,38 @@ func newTypeEnv(f func() *typeChecker) *TypeEnv {
 }
 
 // Get returns the type of x.
+// Deprecated: Use GetByValue or GetByRef instead, as they are more efficient.
 func (env *TypeEnv) Get(x interface{}) types.Type {
-
 	if term, ok := x.(*Term); ok {
 		x = term.Value
 	}
 
-	switch x := x.(type) {
+	if v, ok := x.(Value); ok {
+		return env.GetByValue(v)
+	}
+
+	panic("unreachable")
+}
+
+// GetByValue returns the type of v.
+func (env *TypeEnv) GetByValue(v Value) types.Type {
+	switch x := v.(type) {
 
 	// Scalars.
 	case Null:
-		return types.NewNull()
+		return types.Nl
 	case Boolean:
-		return types.NewBoolean()
+		return types.B
 	case Number:
-		return types.NewNumber()
+		return types.N
 	case String:
-		return types.NewString()
+		return types.S
 
 	// Composites.
 	case *Array:
 		static := make([]types.Type, x.Len())
 		for i := range static {
-			tpe := env.Get(x.Elem(i).Value)
+			tpe := env.GetByValue(x.Elem(i).Value)
 			static[i] = tpe
 		}
 
@@ -63,7 +72,7 @@ func (env *TypeEnv) Get(x interface{}) types.Type {
 		return types.NewArray(static, dynamic)
 
 	case *lazyObj:
-		return env.Get(x.force())
+		return env.GetByValue(x.force())
 	case *object:
 		static := []*types.StaticProperty{}
 		var dynamic *types.DynamicProperty
@@ -72,14 +81,14 @@ func (env *TypeEnv) Get(x interface{}) types.Type {
 			if IsConstant(k.Value) {
 				kjson, err := JSON(k.Value)
 				if err == nil {
-					tpe := env.Get(v)
+					tpe := env.GetByValue(v.Value)
 					static = append(static, types.NewStaticProperty(kjson, tpe))
 					return
 				}
 			}
 			// Can't handle it as a static property, fallback to dynamic
-			typeK := env.Get(k.Value)
-			typeV := env.Get(v.Value)
+			typeK := env.GetByValue(k.Value)
+			typeV := env.GetByValue(v.Value)
 			dynamic = types.NewDynamicProperty(typeK, typeV)
 		})
 
@@ -92,8 +101,7 @@ func (env *TypeEnv) Get(x interface{}) types.Type {
 	case Set:
 		var tpe types.Type
 		x.Foreach(func(elem *Term) {
-			other := env.Get(elem.Value)
-			tpe = types.Or(tpe, other)
+			tpe = types.Or(tpe, env.GetByValue(elem.Value))
 		})
 		if tpe == nil {
 			tpe = types.A
@@ -104,47 +112,46 @@ func (env *TypeEnv) Get(x interface{}) types.Type {
 	case *ArrayComprehension:
 		cpy, errs := env.newChecker().CheckBody(env, x.Body)
 		if len(errs) == 0 {
-			return types.NewArray(nil, cpy.Get(x.Term))
+			return types.NewArray(nil, cpy.GetByValue(x.Term.Value))
 		}
 		return nil
 	case *ObjectComprehension:
 		cpy, errs := env.newChecker().CheckBody(env, x.Body)
 		if len(errs) == 0 {
-			return types.NewObject(nil, types.NewDynamicProperty(cpy.Get(x.Key), cpy.Get(x.Value)))
+			return types.NewObject(nil, types.NewDynamicProperty(cpy.GetByValue(x.Key.Value), cpy.GetByValue(x.Value.Value)))
 		}
 		return nil
 	case *SetComprehension:
 		cpy, errs := env.newChecker().CheckBody(env, x.Body)
 		if len(errs) == 0 {
-			return types.NewSet(cpy.Get(x.Term))
+			return types.NewSet(cpy.GetByValue(x.Term.Value))
 		}
 		return nil
 
 	// Refs.
 	case Ref:
-		return env.getRef(x)
+		return env.GetByRef(x)
 
 	// Vars.
 	case Var:
-		if node := env.tree.Child(x); node != nil {
+		if node := env.tree.Child(v); node != nil {
 			return node.Value()
 		}
 		if env.next != nil {
-			return env.next.Get(x)
+			return env.next.GetByValue(v)
 		}
 		return nil
 
 	// Calls.
 	case Call:
 		return nil
-
-	default:
-		panic("unreachable")
 	}
+
+	return env.Get(v)
 }
 
-func (env *TypeEnv) getRef(ref Ref) types.Type {
-
+// GetByRef returns the type of the value referred to by ref.
+func (env *TypeEnv) GetByRef(ref Ref) types.Type {
 	node := env.tree.Child(ref[0].Value)
 	if node == nil {
 		return env.getRefFallback(ref)
@@ -156,7 +163,7 @@ func (env *TypeEnv) getRef(ref Ref) types.Type {
 func (env *TypeEnv) getRefFallback(ref Ref) types.Type {
 
 	if env.next != nil {
-		return env.next.Get(ref)
+		return env.next.GetByRef(ref)
 	}
 
 	if RootDocumentNames.Contains(ref[0]) {
@@ -439,7 +446,7 @@ func insertIntoObject(o *types.Object, path Ref, tpe types.Type, env *TypeEnv) (
 		return o, nil
 	}
 
-	key := env.Get(path[0].Value)
+	key := env.GetByValue(path[0].Value)
 
 	if len(path) == 1 {
 		var dynamicProps *types.DynamicProperty