And make shorthand types boxed types to avoid allocations at runtime.

Below stats show the difference running Regal lint with all rules disabled (to better highlight compilation costs). @srenatus will be pleased by this format, I'm sure.

```
goos: darwin
goarch: arm64
pkg: github.com/styrainc/regal/pkg/linter
cpu: Apple M4 Pro
                       │  main.txt   │              pr.txt               │
                       │   sec/op    │   sec/op     vs base              │
RegalNoEnabledRules-12   190.8m ± 2%   181.8m ± 2%  -4.72% (p=0.002 n=6)

                       │   main.txt   │               pr.txt               │
                       │     B/op     │     B/op      vs base              │
RegalNoEnabledRules-12   474.4Mi ± 0%   471.3Mi ± 0%  -0.65% (p=0.002 n=6)

                       │  main.txt   │              pr.txt               │
                       │  allocs/op  │  allocs/op   vs base              │
RegalNoEnabledRules-12   9.657M ± 0%   9.476M ± 0%  -1.88% (p=0.002 n=6)
```